epymarl/requirements.txt
absl-py
attrs
gym==0.21.0
idna==2.8
imageio
importlib-metadata<5.0
lbforaging
matplotlib
mpyq==0.2.5
numpy
pbr==4.3.0
Pillow
pluggy==0.7.1
portpicker==1.2.0
probscale
protobuf==3.19.5
py==1.6.0
pygame
pyglet<2
pyparsing==2.2.2
pysc2==3.0.0
pytest==4.3.1
python-dateutil
PyYAML==5.3.1
requests
rware
s2clientprotocol==4.10.1.75800.0
sacred
scipy
six
sk-video==1.1.10
snakeviz
tensorboard-logger
torch
torchvision
tornado
urllib3
websocket-client==0.53.0
whichcraft==0.5.2
wrapt
git+https://github.com/oxwhirl/smac.git#egg=SMAC
lbforaging
